Move makeStringAsyncIterator into testHelpers.ts
